Update interfaces and controllers to use DTOs
The Category and Product interfaces and controllers have been updated for operations to use DTOs instead of models. The selected files mode of IProductService.cs was also changed to 100644 and the ProductController is now more optimized. The validation of incoming and outgoing information has been improved as well. Additional dependencies have also been added to provide more functionalities such as AutoMapper and JwtBearer.